Modify the original Congress chart
