--- src/4cbd138614bf115f/files/fe83fd55eb1882e6cc467c43d748360412553c578e05ed7178591a825b639661bcbcbcadf21e8a48741b7a051744246fa3cd2e3fcb1721dd66618febc1ef29ce
@@ -1,5 +1,0 @@
-﻿Stance,Senate,House of Representative
-Pro-crypto,3,17
-Anti-crypto,2,1
-Thinking about crypto,4,10
-No opinion,91,407
