--- src/4cbd138614bf115f/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf-8">
-<title>US Congress Chart</title>
-<link rel="stylesheet" type="text/css" href="./inspector.css">
-<body>
-<script type="module">
-
-import define from "./index.js";
-import {Runtime, Library, Inspector} from "./runtime.js";
-
-const runtime = new Runtime();
-const main = runtime.module(define, Inspector.into(document.body));
-
-</script>
