--- src/4cbd138614bf115f/index.js
@@ -1,1 +1,0 @@
-export {default} from "./4cbd138614bf115f@585.js";
