+++ src/App.js
@@ -13,7 +13,6 @@ import Tree from "./components/chart/Tree";
 import Congress from "./components/chart/Congress";
 import Chinamap from "./components/chart/Map";
 import Candidate from "./components/chart/Candidate";
-import Donut from './components/chart/Donut';
 
 import './App.css';
 import Container from 'react-bootstrap/Container';
@@ -57,7 +56,6 @@ function App(props) {
   const { width, height } = useContainerDimensions(componentRef);
 
   const interactives = [
-    { interactive: <Donut width = {width} height = {height} />, title: "Chinese NPC Deputy Map", description: "Use react and D3 to show the population and the number of NPC deputies of each Chinese province.", link: "https://haoshuai999.github.io/china-data-map/", date: "Aug. 8 2021"},
     { interactive: <Chinamap width = {width}/>, title: "Chinese NPC Deputy Map", description: "Use react and D3 to show the population and the number of NPC deputies of each Chinese province.", link: "https://haoshuai999.github.io/china-data-map/", date: "Aug. 8 2021"},
     { interactive: <Returns width = {width > 500 ? width : width - 80}/>, title: "Crypto Returns Over Time", description: "Fetch the lastest crypto returns data using the notebook.", link: "https://observablehq.com/d/99803321128e165f", date: "Sept. 14 2020"},
     { interactive: <Candidate width = {width}/>, title: "The 2020 Singaporean General Election", description: "I made an interactive bar chart and a world map showing the growth of female political candidates in Singapore.", link: "https://observablehq.com/d/fb5ed6161a8a33b8", date: "Sept. 2 2021"},
@@ -65,7 +63,7 @@ function App(props) {
     { interactive: <Bar width = {width}/>, title: "Vertical Returns Bar Chart", description: "Compare the returns of different crypto assets.", link: "https://observablehq.com/d/576de58dae43c14e", date: "Sept. 20 2020"},
     { interactive: <Heatmap width = {width}/>, title: "BTC Returns Heatmap", description: "Bitcoin performs better on US trading hours than Asian trading hours.", link: "https://observablehq.com/@shuaihaofzny/heatmap-example", date: "Sept. 30 2020"},
     { interactive: <Radial width = {width}/>, title: "Coinbase Growth", description: "Total assets on Coinbase increased nearly 150% in 2021Q1.", link: "https://observablehq.com/d/684a54bcf8262fbd", date: "Apr. 3 2021"},
-    { interactive: <Congress width = {width}/>, title: "Blockchain and Congress", description: "Most lawmakers don't hold any opinion towards blockchain technology.", link: "https://observablehq.com/d/4cbd138614bf115f", date: "Jan. 7 2021"},
+    { interactive: <Congress width = {width} height= {height}/>, title: "United States Congress", description: "I use some dummy data to make this half donut chart. This can also be used for US Congress related charts", link: "https://observablehq.com/d/7c527de234abd676", date: "Jan. 7 2021"},
     { interactive: <Tree width = {width}/>, title: "Crypto Family Tree", description: "Many crypto assets are related to each other through forking.", link: "https://observablehq.com/d/6a0e7e787418caa2", date: "Mar. 21 2021"}
   ]
